+++ src/quick-lint-js/diag/diagnostic-types-2.h
@@ -2367,6 +2367,16 @@ struct Diag_TypeScript_Interfaces_Cannot_Contain_Static_Blocks {
   Source_Code_Span static_token;
 };
 
+struct Diag_TypeScript_Generic_Less_Less_Not_Split {
+  [[qljs::diag("E0429", Diagnostic_Severity::error)]]  //
+  // clang-format off
+  [[qljs::message("space is required between '<' and '<' inside 'extends' "
+                  "or 'implements' clause",
+                  ARG(expected_space))]]  //
+  // clang-format on
+  Source_Code_Span expected_space;
+};
+
 struct Diag_TypeScript_Declare_Class_Cannot_Contain_Static_Block_Statement {
   [[qljs::diag("E0332", Diagnostic_Severity::error)]]  //
   [[qljs::message("'declare class' cannot contain static block",