+++ src/quick-lint-js/i18n/translation-table-generated.h
@@ -18,8 +18,8 @@ namespace quick_lint_js {
 using namespace std::literals::string_view_literals;
 
 constexpr std::uint32_t translation_table_locale_count = 5;
-constexpr std::uint16_t translation_table_mapping_table_size = 532;
-constexpr std::size_t translation_table_string_table_size = 80499;
+constexpr std::uint16_t translation_table_mapping_table_size = 533;
+constexpr std::size_t translation_table_string_table_size = 80577;
 constexpr std::size_t translation_table_locale_table_size = 35;
 
 QLJS_CONSTEVAL std::uint16_t translation_table_const_look_up(
@@ -471,6 +471,7 @@ QLJS_CONSTEVAL std::uint16_t translation_table_const_look_up(
           "see here"sv,
           "semicolon is not allowed after decorators"sv,
           "something happened"sv,
+          "space is required between '<' and '<' inside 'extends' or 'implements' clause"sv,
           "spread starts here"sv,
           "spread tuple elements cannot be optional"sv,
           "static block starts here"sv,